+++ 0x0A-argc_argv/holberton.h
@@ -2,6 +2,7 @@
 #define HOLBERTON
 
 int _putchar(char c);
-int atoi(char *);
+long int atoi(char *);
+
 
 #endif
